requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
pydantic==2.11.7
email-validator==2.2.0
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
google-generativeai==0.3.2
psycopg2-binary